refactor(gsd): narrow lifecycle helper test seams
