+++ src/resources/extensions/gsd/auto-worktree-path-resolution.ts
@@ -1,0 +1,6 @@
+// gsd-pi — Narrow auto-worktree path resolution seam.
+//
+// Keeps callers that only resolve existing auto-worktree paths off the legacy
+// auto-worktree compatibility barrel while extraction continues.
+
+export { getAutoWorktreePath } from "./auto-worktree.js";
